CloakDocs-Backend/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
flask-restx==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
Werkzeug==3.1.3
# PDF İşleme ve Anahtar Kelime Çıkarımı
PyPDF2==3.0.1
pdfminer.six==20221105
spacy==3.7.2
scikit-learn==1.3.2
keybert==0.8.3
yake==0.4.8
